+++ py-polars/tests/unit/utils/test_utils.py
@@ -23,7 +23,6 @@ from polars._utils.various import (
     parse_percentiles,
     parse_version,
 )
-from polars.io._utils import looks_like_url
 
 if TYPE_CHECKING:
     from zoneinfo import ZoneInfo
@@ -291,22 +290,3 @@ def test_is_str_sequence_check(
     assert is_str_sequence(sequence, include_series=include_series) == expected
     if expected:
         assert is_sequence(sequence, include_series=include_series)
-
-
-@pytest.mark.parametrize(
-    ("url", "result"),
-    [
-        ("HTTPS://pola.rs/data.csv", True),
-        ("http://pola.rs/data.csv", True),
-        ("ftps://pola.rs/data.csv", True),
-        ("FTP://pola.rs/data.csv", True),
-        ("htp://pola.rs/data.csv", False),
-        ("fttp://pola.rs/data.csv", False),
-        ("http_not_a_url", False),
-        ("ftp_not_a_url", False),
-        ("/mnt/data.csv", False),
-        ("file://mnt/data.csv", False),
-    ],
-)
-def test_looks_like_url(url: str, result: bool) -> None:
-    assert looks_like_url(url) == result
